use one FakeSite; fix bugs in markdown plugin magic
